cleanup and better component for example
