+++ examples/with-ai-sdk/app/MyRuntimeProvider.tsx
@@ -1,40 +1,15 @@
 "use client";
 
-import {
-  AssistantRuntimeProvider,
-  makeAssistantTool,
-  tool,
-  useAssistantRuntime,
-  // makeAssistantTool,
-  // makeAssistantToolUI,
-  // tool,
-} from "@assistant-ui/react";
+import { AssistantRuntimeProvider } from "@assistant-ui/react";
 import { useVercelUseChatRuntime } from "@assistant-ui/react-ai-sdk";
 import { useChat } from "@ai-sdk/react";
-import { Tools } from "@/lib/tools-2";
+import { Tools } from "@/lib/tools";
 import { assistantUIToolBox } from "@assistant-ui/react/runtimes/core/ThreadRuntimeCore";
-// import { z } from "zod";
 
 export const toolbox = assistantUIToolBox<Tools>();
 
-// const Test = toolbox.getLocationFromUser.getUI({
-//   execute: (a) => "test",
-//   render: ({ result }) => {
-//     console.log("result: ", result);
-//     return <div>test</div>;
-//   },
-// });
-
-const Test = toolbox.getLocationFromUser.getTool({
-  // execute: async () => {
-  //   console.log("execute: ");
-  //   return "Mumbai";
-  // },
+const GetLocationFromUser = toolbox.getLocationFromUser.getTool({
   render: (a) => {
-    // console.log("result: ", a.result);
-    // console.log("result: ", result);
-
-    // return <div>{a.result}</div>;
     return (
       <button
         onClick={() => {
@@ -47,55 +22,26 @@ const Test = toolbox.getLocationFromUser.getTool({
   },
 });
 
-// const getLocation = tool({
-//   parameters: z.object({
-//     location: z.string(),
-//   }),
-//   execute: async ({ location }) => {
-//     console.log("execute: ", location);
-//     return location;
-//   },
-// });
-
-// const Test = makeAssistantTool({
-//   toolName: "weather",
-//   parameters: z.object({
-//     location: z.string(),
-//   }),
-//   execute: async ({ location }) => {
-//     console.log("execute: ", location);
-//     return 100;
-//   },
-//   render: ({ result }) => {
-//     console.log("result: ", result);
-//     return <div>{result}</div>;
-//   },
-// });
-
-// const submitForm = tool({
-//   parameters: z.object({
-//     email: z.string().email(),
-//     name: z.string(),
-//   }),
-//   execute: async ({ email, name }) => {
-//     // Implementation
-//     return { success: true };
-//   },
-// });
-
-// // // Create a tool component
-// const SubmitFormTool = makeAssistantTool({
-//   toolName: "submitForm",
-//   ...submitForm,
-// });
-
-// const WTF = makeAssistantToolUI({
-//   toolName: "weather",
-//   render: ({ result }) => {
-//     console.log("result: ", result);
-//     return <div>test</div>;
-//   },
-// });
+const Weather = toolbox.weather.getTool({
+  render: ({ result }) => {
+    return (
+      <div className="mx-auto flex max-w-sm items-center space-x-4 rounded-xl bg-white p-6 shadow-lg">
+        <div className="shrink-0">
+          {/* Weather icon - using a simple emoji as placeholder */}
+          <span className="text-4xl">☀️</span>
+        </div>
+        <div>
+          <div className="text-xl font-medium text-black">Weather Report</div>
+          <p className="text-slate-500">
+            The weather in{" "}
+            <span className="font-semibold text-blue-600">{result}</span> is{" "}
+            <span className="font-bold">20°C</span>
+          </p>
+        </div>
+      </div>
+    );
+  },
+});
 
 export function MyRuntimeProvider({
   children,
@@ -106,23 +52,10 @@ export function MyRuntimeProvider({
 
   const runtime = useVercelUseChatRuntime(chat);
 
-  console.log("yurr: ", chat.messages);
-
-  // console.log(
-  //   "wtf: ",
-  //   toolbox.getLocationFromUser.getUI({
-  //     execute: (a) => "test",
-  //     render: ({ result }) => {
-  //       console.log("result: ", result);
-  //       return <div>test</div>;
-  //     },
-  //   }),
-  // );
-
   return (
     <AssistantRuntimeProvider runtime={runtime}>
-      <Test />
-      {/* <WTF /> */}
+      <GetLocationFromUser />
+      <Weather />
       {children}
     </AssistantRuntimeProvider>
   );
